+++ test/format/test_json.py
@@ -5,15 +5,15 @@ import pytest
 import pip_audit._format as format
 
 
-@pytest.mark.parametrize("output_desc", [True, False])
-def test_json_manifest(output_desc):
-    fmt = format.JsonFormat(output_desc)
+@pytest.mark.parametrize("output_desc, output_aliases", ([True, False], [True, False]))
+def test_json_manifest(output_desc, output_aliases):
+    fmt = format.JsonFormat(output_desc, output_aliases)
 
     assert fmt.is_manifest
 
 
 def test_json(vuln_data):
-    json_format = format.JsonFormat(True)
+    json_format = format.JsonFormat(True, True)
     expected_json = {
         "dependencies": [
             {
@@ -26,11 +26,13 @@ def test_json(vuln_data):
                             "1.1",
                             "1.4",
                         ],
+                        "aliases": ["CVE-0000-00000"],
                         "description": "The first vulnerability",
                     },
                     {
                         "id": "VULN-1",
                         "fix_versions": ["1.0"],
+                        "aliases": ["CVE-0000-00001"],
                         "description": "The second vulnerability",
                     },
                 ],
@@ -42,6 +44,7 @@ def test_json(vuln_data):
                     {
                         "id": "VULN-2",
                         "fix_versions": [],
+                        "aliases": ["CVE-0000-00002"],
                         "description": "The third vulnerability",
                     }
                 ],
@@ -53,7 +56,47 @@ def test_json(vuln_data):
 
 
 def test_json_no_desc(vuln_data):
-    json_format = format.JsonFormat(False)
+    json_format = format.JsonFormat(False, True)
+    expected_json = {
+        "dependencies": [
+            {
+                "name": "foo",
+                "version": "1.0",
+                "vulns": [
+                    {
+                        "id": "VULN-0",
+                        "fix_versions": [
+                            "1.1",
+                            "1.4",
+                        ],
+                        "aliases": ["CVE-0000-00000"],
+                    },
+                    {
+                        "id": "VULN-1",
+                        "fix_versions": ["1.0"],
+                        "aliases": ["CVE-0000-00001"],
+                    },
+                ],
+            },
+            {
+                "name": "bar",
+                "version": "0.1",
+                "vulns": [
+                    {
+                        "id": "VULN-2",
+                        "fix_versions": [],
+                        "aliases": ["CVE-0000-00002"],
+                    }
+                ],
+            },
+        ],
+        "fixes": [],
+    }
+    assert json_format.format(vuln_data, list()) == json.dumps(expected_json)
+
+
+def test_json_no_desc_no_aliases(vuln_data):
+    json_format = format.JsonFormat(False, False)
     expected_json = {
         "dependencies": [
             {
@@ -76,7 +119,12 @@ def test_json_no_desc(vuln_data):
             {
                 "name": "bar",
                 "version": "0.1",
-                "vulns": [{"id": "VULN-2", "fix_versions": []}],
+                "vulns": [
+                    {
+                        "id": "VULN-2",
+                        "fix_versions": [],
+                    }
+                ],
             },
         ],
         "fixes": [],
@@ -85,7 +133,7 @@ def test_json_no_desc(vuln_data):
 
 
 def test_json_skipped_dep(vuln_data_skipped_dep):
-    json_format = format.JsonFormat(False)
+    json_format = format.JsonFormat(False, True)
     expected_json = {
         "dependencies": [
             {
@@ -98,6 +146,7 @@ def test_json_skipped_dep(vuln_data_skipped_dep):
                             "1.1",
                             "1.4",
                         ],
+                        "aliases": ["CVE-0000-00000"],
                     },
                 ],
             },
@@ -112,7 +161,7 @@ def test_json_skipped_dep(vuln_data_skipped_dep):
 
 
 def test_json_fix(vuln_data, fix_data):
-    json_format = format.JsonFormat(True)
+    json_format = format.JsonFormat(True, True)
     expected_json = {
         "dependencies": [
             {
@@ -125,11 +174,13 @@ def test_json_fix(vuln_data, fix_data):
                             "1.1",
                             "1.4",
                         ],
+                        "aliases": ["CVE-0000-00000"],
                         "description": "The first vulnerability",
                     },
                     {
                         "id": "VULN-1",
                         "fix_versions": ["1.0"],
+                        "aliases": ["CVE-0000-00001"],
                         "description": "The second vulnerability",
                     },
                 ],
@@ -141,6 +192,7 @@ def test_json_fix(vuln_data, fix_data):
                     {
                         "id": "VULN-2",
                         "fix_versions": [],
+                        "aliases": ["CVE-0000-00002"],
                         "description": "The third vulnerability",
                     }
                 ],
@@ -163,7 +215,7 @@ def test_json_fix(vuln_data, fix_data):
 
 
 def test_json_skipped_fix(vuln_data, skipped_fix_data):
-    json_format = format.JsonFormat(True)
+    json_format = format.JsonFormat(True, True)
     expected_json = {
         "dependencies": [
             {
@@ -176,11 +228,13 @@ def test_json_skipped_fix(vuln_data, skipped_fix_data):
                             "1.1",
                             "1.4",
                         ],
+                        "aliases": ["CVE-0000-00000"],
                         "description": "The first vulnerability",
                     },
                     {
                         "id": "VULN-1",
                         "fix_versions": ["1.0"],
+                        "aliases": ["CVE-0000-00001"],
                         "description": "The second vulnerability",
                     },
                 ],
@@ -192,6 +246,7 @@ def test_json_skipped_fix(vuln_data, skipped_fix_data):
                     {
                         "id": "VULN-2",
                         "fix_versions": [],
+                        "aliases": ["CVE-0000-00002"],
                         "description": "The third vulnerability",
                     }
                 ],